Move token_stream to separate module
